:ambulance: fix swagger
