+++ topping/package.json
@@ -76,8 +76,7 @@
 		"build": "rimraf ./dist & tsc",
 		"watch": "tsc -w",
 		"gen-gql": "rimraf ./modules/graphql & gqlg --schemaFilePath ./src/schema.graphql --destDirPath ./src/modules/graphql --depthLimit 5",
-		"gen-env": "npx gen-env-types .env -o src/types/env.d.ts -e .",
-		"gen-rest-api-docs": "ts-node ./src/scripts/genOpenAPI.ts"
+		"gen-env": "npx gen-env-types .env -o src/types/env.d.ts -e ."
 	},
 	"resolutions": {
 		"graphql": "^15.0.0"
