+++ topping/src/modules/resolvers/user/get_user/get_user.dto.ts
@@ -1,11 +1,6 @@
 import { Field, InputType } from "type-graphql";
-import * as yup from "yup";
 @InputType()
 export class GetUserDto {
 	@Field()
 	userId: string;
 }
-
-export const YUP_USER_READ = yup.object().shape({
-	userId: yup.string().uuid(),
-});
