+++ topping/src/modules/resolvers/user/get_user/get_user.validate.ts
@@ -1,0 +1,5 @@
+import * as yup from "yup";
+
+export const YUP_USER_READ = yup.object().shape({
+	userId: yup.string().uuid(),
+});
