+++ topping/src/modules/resolvers/user/get_users/get_users.resolver.ts
@@ -2,7 +2,6 @@ import { Arg, Resolver, Query, UseMiddleware } from "type-graphql";
 import { InjectRepository } from "typeorm-typedi-extensions";
 import { yupValidateMiddleware } from "../../../middleware/yupValidate";
 import { UserRepository } from "../../../repository/user/UserRepository";
-import { GetUserDto, YUP_USER_READ } from "../get_user/get_user.dto";
 import { User } from "../../../../entity/User";
 
 @Resolver((of) => User)
