+++ topping/src/modules/resolvers/user/login/login.dto.ts
@@ -1,6 +1,4 @@
 import { Field, InputType } from "type-graphql";
-import * as yup from "yup";
-import { sharedSchema } from "../../../../shared/yupSchema";
 
 @InputType()
 export class LoginDto {
@@ -10,8 +8,3 @@ export class LoginDto {
 	@Field()
 	password: string;
 }
-
-export const YUP_LOGIN = yup.object().shape({
-	email: sharedSchema.email,
-	password: sharedSchema.password,
-});
