+++ topping/src/modules/resolvers/user/login/login.validate.ts
@@ -1,0 +1,7 @@
+import * as yup from "yup";
+import { sharedSchema } from "../../../../shared/yupSchema";
+
+export const YUP_LOGIN = yup.object().shape({
+	email: sharedSchema.email,
+	password: sharedSchema.password,
+});
