+++ topping/src/modules/resolvers/user/register/register.dto.ts
@@ -1,6 +1,4 @@
 import { Field, InputType } from "type-graphql";
-import * as yup from "yup";
-import { sharedSchema } from "../../../../shared/yupSchema";
 
 @InputType()
 export class RegisterDto {
@@ -19,11 +17,3 @@ export class RegisterDto {
 	@Field()
 	password: string;
 }
-
-export const YUP_REGISTER = yup.object().shape({
-	firstName: yup.string().min(3).max(255),
-	lastName: yup.string().min(3).max(255),
-	email: sharedSchema.email,
-	password: sharedSchema.password,
-	username: yup.string().min(4).max(20),
-});
